requirements.txt
streamlit
pdfplumber
python-dotenv
Flask 
requests


